requirements.txt
asgiref==3.2.10
bleach==3.1.1
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
cffi==1.14.0
cryptography==2.8
Django==3.0.8
django-jsonfield==1.4.0
django-randomslugfield==0.3.0
django-storages==1.9.1
docutils==0.15.2
importlib-metadata==1.6.1
jeepney==0.4.2
jmespath==0.10.0
keyring==21.1.0
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.5.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
readme-renderer==24.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
SecretStorage==3.1.2
six==1.15.0
sqlparse==0.3.1
twine==3.1.1
urllib3==1.25.9
webencodings==0.5.1
zipp==3.1.0
